Fix build errors in native
